# apps/backend/internal/adapter/httpapi/generated/openapi.codegen.yaml
openapi: 3.0.3
info:
  title: Telomoyo POS API
  version: 1.0.0
  description: |
    The contract used by the Android terminal and the Go backend. Money is integer
    Indonesian rupiah, timestamps are RFC 3339, reporting uses Asia/Jakarta, and
    transaction IDs are raw ULIDs. User-facing surfaces add the `TRX-` prefix.

    Unless a response is a downloadable export, successful JSON responses use
    `{data, meta}` and errors use `{error: {code, message, details, requestId}}`.
servers:
  - url: /api/v1
    description: Version 1 API
security:
  - bearerAuth: []
...
paths:
  /auth/contexts:
    get:
      operationId: getAuthContexts
      tags: [Authentication]
      summary: List the account's available tenant and platform contexts
      responses:
        "200":
          description: List the account's available tenant and platform contexts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableContextsEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /auth/switch-context:
    post:
      operationId: switchAuthContext
      tags: [Authentication]
      summary: Exchange the session into an authorized account, tenant, or platform context
      description: Tenant context always starts in Production. Requires online outbox drainage or explicit quarantine after revoked access. The previous session is revoked and retained as immutable origin evidence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchContextRequest"
      responses:
        "200":
          description: Exchange the session into an authorized account, tenant, or platform context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /auth/register-invitation:
    post:
      operationId: registerInvitation
      tags: [Authentication]
      summary: Register an account and atomically accept a valid invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInvitationRequest"
      responses:
        "201":
          description: Register an account and atomically accept a valid invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /auth/invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [Authentication]
      summary: Accept an invitation using the authenticated account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationRequest"
      responses:
        "200":
          description: Accept an invitation using the authenticated account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantContextEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /platform/tenants:
    get:
      operationId: listPlatformTenants
      tags: [Platform]
      summary: List tenants in platform context
      responses:
        "200":
          description: List tenants in platform context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    post:
      operationId: createTenant
      tags: [Platform]
      summary: Create a pending tenant and a seven-day initial-owner invitation
      description: Requires platform context and TENANT_PROVISIONING_ENABLED. Does not grant tenant membership to the platform administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Create a pending tenant and a seven-day initial-owner invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTenantEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /platform/tenants/{tenantId}/status:
    post:
      operationId: setTenantStatus
      tags: [Platform]
      summary: Suspend or reactivate a tenant without reviving revoked sessions
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetTenantStatusRequest"
      responses:
        "200":
          description: Suspend or reactivate a tenant without reviving revoked sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /platform/tenants/{tenantId}/invitation:
    post:
      operationId: reissueOwnerInvitation
      tags: [Platform]
      summary: Reissue the initial-owner invitation for a pending tenant
      description: Invalidates the previous initial-owner code. Active tenant membership is managed only by that tenant's superadmins.
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Reissue the initial-owner invitation for a pending tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /platform/audit:
    get:
      operationId: listPlatformAudit
      tags: [Platform]
      summary: List recent durable control-plane audit events
      parameters:
        - $ref: "#/components/parameters/AuditLimit"
      responses:
        "200":
          description: List recent durable control-plane audit events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformAuditListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /tenant/members:
    get:
      operationId: listTenantMembers
      tags: [Tenants]
      summary: List memberships in the active tenant
      responses:
        "200":
          description: List memberships in the active tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMemberListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /tenant/members/{userId}:
    patch:
      operationId: updateTenantMember
      tags: [Tenants]
      summary: Update the role or active status of one tenant membership
      description: Production-mode tenant superadmin only. Does not change global credentials or other tenant memberships; last-active-superadmin and self-demotion protections apply.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantMemberRequest"
      responses:
        "200":
          description: Update the role or active status of one tenant membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMemberEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /tenant/invitations:
    get:
      operationId: listTenantInvitations
      tags: [Tenants]
      summary: List tenant invitations without exposing their codes
      responses:
        "200":
          description: List tenant invitations without exposing their codes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    post:
      operationId: createTenantInvitation
      tags: [Tenants]
      summary: Create a single-use seven-day membership invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: Create a single-use seven-day membership invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /tenant/invitations/{invitationId}:
    delete:
      operationId: revokeTenantInvitation
      tags: [Tenants]
      summary: Revoke an unused tenant invitation
      parameters:
        - $ref: "#/components/parameters/InvitationId"
      responses:
        "200":
          description: Revoke an unused tenant invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /tenant/profile:
    get:
      operationId: getTenantProfile
      tags: [Tenants]
      summary: Get the current business receipt and report identity
      responses:
        "200":
          description: Get the current business receipt and report identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantProfileEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    patch:
      operationId: updateTenantProfile
      tags: [Tenants]
      summary: Create a new tenant business-profile revision
      description: Production-mode tenant superadmin only. Existing transactions retain their captured identity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantProfileRequest"
      responses:
        "200":
          description: Create a new tenant business-profile revision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantProfileEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /tenant/qris:
    get:
      operationId: getTenantQRIS
      tags: [Tenants]
      summary: Get active and historical merchant QRIS payloads for the tenant
      responses:
        "200":
          description: Get active and historical merchant QRIS payloads for the tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantQRISConfigEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    put:
      operationId: updateTenantQRIS
      tags: [Tenants]
      summary: Import a merchant QRIS payload and optionally activate it
      description: Production-mode tenant superadmin only. Historical payloads remain available for exact transaction bindings. activate defaults to true; false imports only a historical payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantQRISRequest"
      responses:
        "200":
          description: Import a merchant QRIS payload and optionally activate it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantQRISConfigEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /auth/logout:
    post:
      operationId: logout
      tags: [Authentication]
      summary: Revoke the current session
      description: The terminal must drain its outbox before calling this online-only endpoint.
      responses:
        "200":
          $ref: "#/components/responses/Action"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /sandbox/status:
    get:
      operationId: getSandboxStatus
      tags: [Sandbox]
      summary: Get Sandbox availability and active generation
      responses:
        "200":
          $ref: "#/components/responses/SandboxStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpaqueSessionToken
      description: Opaque 256-bit token. Only its hash is persisted by the backend.
...
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
...
    BadRequest:
      $ref: "#/components/responses/ErrorResponse"
    Unauthorized:
      $ref: "#/components/responses/ErrorResponse"
    Forbidden:
      $ref: "#/components/responses/ErrorResponse"
    NotFound:
      $ref: "#/components/responses/ErrorResponse"
    Conflict:
      $ref: "#/components/responses/ErrorResponse"
...
    UnprocessableEntity:
      $ref: "#/components/responses/ErrorResponse"
...
  schemas:
...
    AvailableContextsEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AvailableContexts"
        meta:
          $ref: "#/components/schemas/Meta"
...
    UUID:
      type: string
      format: uuid
...